+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+### Added
+- A single way to calculate the final time.
+### Fixed
+- Time step calculation for Runge Kutta Forth Order method.
+
 ## [2.0.1] - 2020-05-07
 ### Fixed
 - CalculateVibration_RigidBody derived classes constructor.